# Add liballoc impl SpecFromElem for i8