@pickfire 2020-06-27
> I don't know offhand how to get the compiler to generate a rep stosw (or d or q). (And in any case, it would need benchmarking to ensure that it's actually faster.)
> I don't know if that would be a performance win, though it might be worth benchmarking (via a cfg-guarded specialization).

I believe we should add some TODO there from @joshtriplett for the other stuff but since I am not sure how to get those working.